feature/shiftgen/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
discord.py>=2.3.0
certifi>=2023.0.0
pytz>=2023.3